Filters instead of macros for BiGG ID styling. Improved handling of model-specific metabolites.
